implement roundRobin library function
This commit implements round-robining between streams within a finite
Traversable container. The testcase adds checks for accurate results,
and presumably termination with CI timeouts, on a finite container of
infinite streams. A minor version bump with changelogging is done to
reflect an addition to an exposed API so that dependencies on it can
be enforced by callers.